Move files to standard locations
[#152464345]
